Avoid duplicate symbols when including perm16.hpp in multiple translation units
The purpose of this commit is to avoid duplicate symbols at link time,
when perm16.hpp is included in multiple implementation (*.cc/.cpp) files.
Without these changes (or some equivalent changes) it is impossible to
include HPCombi in libsemigroups because linking fails because of
duplicate symbols for the static data members and member functions.