Rename VERSION to .VERSION
The reason for this change is that "version" is a C++ standard library
header file from C++20 onwards, and the file VERSION can be confused
with this header when trying to "#include <version>". Even if there are
no explicit "#include <version>" in HPCombi, some standard library files
include "version" and then fail to compile if there's another file named
"VERSION" in the include path.