Temporarily use ubuntu-22.04 instead of latest in CI
This is due to ubuntu-latest being updated from ubuntu-22.04 to
ubuntu-24.04 and there being different versions of gcc/clang available
in ubuntu-24.04 and ubuntu-22.04. See the following for some details.

https://github.com/actions/runner-images/issues/10636

A proper fix to this would be one which installs the relevant versions
of gcc/clang in which ever version of ubuntu the runner is using. It's
not completely clear to me how to do that at the moment.