Remove shadow declarations
This commit renames some of the variables from the constructors of
PTransf16 and its derived classes from "v" to "vv". The reason for doing
this is that there is a data member of Vect16 named "v", and compiling
with warnings enabled complains that the variables "v" shadowed the data
member "v".